+++ src/lib/data/navigation-data.ts
@@ -6,6 +6,7 @@ export const bottomNavigation = [
   { name: 'Information', href: '#' },
   { name: 'Vacancies', href: '#' },
   { name: 'Contact Us', href: '/contactus' },
+  { name: 'Test Page', href: '/testfolder' },
 ];
 export const navigation = {
   categories: [
